Inline functions from shallow utility modules, and rename them to the selectorFor_* pattern
